fix some warnings

+++ GtraceLib/FileUtils.cpp
@@ -38,7 +38,7 @@ int readLine(char* buffer, int n, FILE* fp) {
 	}
 	// remove the new line symbol
 	if (res[size - 1] == '\n') res[--size] = '\0';
-	return size;
+	return static_cast<int>(size);
 }
 
 }
